refactor: Separate locust workloads into different files
This removes the need for tags and makes the start command simpler.
We still can combine the workloads by passing multiple locustfiles
during launch as in:
```
locust -H 128.0.0.1:3030 \
     -f locustfiles/ft.py,locustfiles/social.py \
     --fungible-token-wasm=$FT_CONTRACT \
     --social-db-wasm=$SOCIAL_DB_CONTRACT \
     --funding-key=$KEY \
     --headless
```
