+++ pytest/tests/loadtest/locust/README.md
@@ -31,10 +31,9 @@ Then to actually run it, this is the command. (Update ports and IP according to 
 ```sh
 cd pytest/tests/loadtest/locust/
 locust -H 127.0.0.1:3030 \
-  FTTransferUser \
+  -f locustfiles/ft.py \
   --fungible-token-wasm=$CONTRACT \
-  --funding-key=$KEY \
-  --tags ft
+  --funding-key=$KEY
 ```
 
 This will print a link to a web UI where the loadtest can be started and statistics & graphs can be observed.
@@ -56,10 +55,9 @@ will work automagically.)
 Start the master
 ```sh
 locust -H 127.0.0.1:3030 \
-  FTTransferUser \
+  -f locustfiles/ft.py \
   --fungible-token-wasm=$CONTRACT \
   --funding-key=$KEY \
-  --tags ft \
   --master
 ```
 
@@ -70,10 +68,9 @@ On the worker
 ulimit -S -n 100000
 # Run the worker, key must include the same account id as used on master
 locust -H 127.0.0.1:3030 \
-  FTTransferUser \
+  -f locustfiles/ft.py \
   --fungible-token-wasm=$CONTRACT \
   --funding-key=$KEY \
-  --tags ft \
   --worker
 ```
 
@@ -82,10 +79,9 @@ Spawning N workers in a single shell:
 for i in {1..16}
 do
    locust -H 127.0.0.1:3030 \
-    FTTransferUser \
+    -f locustfiles/ft.py \
     --fungible-token-wasm=$CONTRACT \
     --funding-key=$KEY \
-    --tags ft \
     --worker &
 done
 ```
@@ -95,22 +91,16 @@ Stopping the master will also terminate all workers.
 
 # Available load types
 
-User classes can be used as subcommand of `locust` to run a specific load test.
-By default, all user classes are combined. But the example above selects
-specifically `FTTransferUser`.
-
-
-When selecting user classes, it is also necessary to select the right tags that
-control initialization code. For example, running with `locustFTTransferUser`
-requires `--tags ft`.
+Different locust files can be passed as an argument to `locust` to run a specific load test.
+All available workloads can be found in `locustfiles` folder.
 
 Currently supported load types:
 
-| load type | user classes | tag | args | description |
+| load type | file | args | description |
 |---|---|---|---|---|
-| Fungible Token | `FTTransferUser` | `ft` | `--fungible-token-wasm $WASM_PATH` <br> (`--num-ft-contracts $N`) |  Creates `$N` FT contracts per worker, registers each user in one of them. Users transfer FTs between each other. |
-| Social DB  | `SocialDbUser` | `social` | `--social-db-wasm $WASM_PATH` | Creates a single instance of SocialDB and registers users to it. Users post messages and follow other users. (More workload TBD) |
-| Congestion | `CongestionUser` | `congestion` | `--congestion-wasm $WASM_PATH` | Creates a single instance of Congestion contract. Users run large and long transactions. |
+| Fungible Token | ft.py | `--fungible-token-wasm $WASM_PATH` <br> (`--num-ft-contracts $N`) |  Creates `$N` FT contracts per worker, registers each user in one of them. Users transfer FTs between each other. |
+| Social DB  | social.py | `--social-db-wasm $WASM_PATH` | Creates a single instance of SocialDB and registers users to it. Users post messages and follow other users. (More workload TBD) |
+| Congestion | congestion.py | `--congestion-wasm $WASM_PATH` | Creates a single instance of Congestion contract. Users run large and long transactions. |
 
 For the congestion workload you will need to use the following contract:
